# IGNITE-23707 Optimize update() in hybrid clocks