--- a/modules/core/src/test/java/org/apache/ignite/internal/hlc/benchmarks/HybridClockBenchmark.java
+++ b/modules/core/src/test/java/org/apache/ignite/internal/hlc/benchmarks/HybridClockBenchmark.java
@@ -17,9 +17,12 @@
 
 package org.apache.ignite.internal.hlc.benchmarks;
 
+import static org.apache.ignite.internal.hlc.HybridTimestamp.LOGICAL_TIME_BITS_SIZE;
+
 import java.util.concurrent.TimeUnit;
 import org.apache.ignite.internal.hlc.HybridClock;
 import org.apache.ignite.internal.hlc.HybridClockImpl;
+import org.apache.ignite.internal.hlc.HybridTimestamp;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Fork;
@@ -75,9 +78,33 @@ public class HybridClockBenchmark {
         hybridClockNow();
     }
 
+    @Benchmark
+    @Threads(1)
+    public void hybridClockUpdateSingleThread() {
+        hybridClockUpdate();
+    }
+
+    @Benchmark
+    @Threads(5)
+    public void hybridClockUpdateFiveThreads() {
+        hybridClockUpdate();
+    }
+
+    @Benchmark
+    @Threads(10)
+    public void hybridClockUpdateTenThreads() {
+        hybridClockUpdate();
+    }
+
     private void hybridClockNow() {
         for (int i = 0; i < 1000; i++) {
             clock.now();
+        }
+    }
+
+    private void hybridClockUpdate() {
+        for (int i = 0; i < 1000; i++) {
+            clock.update(HybridTimestamp.hybridTimestamp((System.currentTimeMillis() << LOGICAL_TIME_BITS_SIZE) + i));
         }
     }