--- a/modules/core/src/test/java/org/apache/ignite/internal/hlc/benchmarks/HybridClockBenchmark.java
+++ b/modules/core/src/test/java/org/apache/ignite/internal/hlc/benchmarks/HybridClockBenchmark.java
@@ -46,8 +46,8 @@ import org.openjdk.jmh.runner.options.OptionsBuilder;
 @Fork(1)
 @Warmup(iterations = 5, time = 2)
 @Measurement(iterations = 10, time = 2)
-@BenchmarkMode(Mode.AverageTime)
-@OutputTimeUnit(TimeUnit.MICROSECONDS)
+@BenchmarkMode(Mode.Throughput)
+@OutputTimeUnit(TimeUnit.SECONDS)
 public class HybridClockBenchmark {
     /** Clock to benchmark. */
     private HybridClock clock;
@@ -104,7 +104,9 @@ public class HybridClockBenchmark {
 
     private void hybridClockUpdate() {
         for (int i = 0; i < 1000; i++) {
-            clock.update(HybridTimestamp.hybridTimestamp((System.currentTimeMillis() << LOGICAL_TIME_BITS_SIZE) + i));
+            long ts = ((System.currentTimeMillis() + (i % 2 == 0 ? 5 : -5)) << LOGICAL_TIME_BITS_SIZE) + (i % 10);
+
+            clock.update(HybridTimestamp.hybridTimestamp(ts));
         }
     }