--- a/modules/transactions/src/main/java/org/apache/ignite/internal/tx/impl/HeapLockManager.java
+++ b/modules/transactions/src/main/java/org/apache/ignite/internal/tx/impl/HeapLockManager.java
@@ -174,7 +174,7 @@ public class HeapLockManager extends AbstractEventProducer<LockEvent, LockEventP
         if (lockKey.contextId() == null) { // Treat this lock as a hierarchy(coarse) lock.
             CoarseLockState state = coarseMap.computeIfAbsent(lockKey, key -> new CoarseLockState(lockKey));
 
-            return state.acquire(txId, lockKey, lockMode);
+            return state.acquire(txId, lockMode);
         }
 
         while (true) {
@@ -521,11 +521,10 @@ public class HeapLockManager extends AbstractEventProducer<LockEvent, LockEventP
          * Acquires a lock.
          *
          * @param txId Tx id.
-         * @param lockKey Lock key.
          * @param lockMode Lock mode.
          * @return The future.
          */
-        public CompletableFuture<Lock> acquire(UUID txId, LockKey lockKey, LockMode lockMode) {
+        public CompletableFuture<Lock> acquire(UUID txId, LockMode lockMode) {
             switch (lockMode) {
                 case S:
                     stripedLock.writeLock().lock();