```
Before
Benchmark                                            Mode  Cnt       Score      Error  Units
HybridClockBenchmark.hybridClockNowFiveThreads      thrpt   10   37536.317 ± 2033.320  ops/s
HybridClockBenchmark.hybridClockNowSingleThread     thrpt   10  238352.974 ± 4753.761  ops/s
HybridClockBenchmark.hybridClockNowTenThreads       thrpt   10   34136.852 ±  283.247  ops/s
HybridClockBenchmark.hybridClockUpdateFiveThreads   thrpt   10   14143.705 ±  365.301  ops/s
HybridClockBenchmark.hybridClockUpdateSingleThread  thrpt   10  124273.226 ± 3827.199  ops/s
HybridClockBenchmark.hybridClockUpdateTenThreads    thrpt   10    9028.719 ±  249.099  ops/s


After
Benchmark                                            Mode  Cnt       Score      Error  Units
HybridClockBenchmark.hybridClockNowFiveThreads      thrpt   10   35338.271 ± 5234.633  ops/s
HybridClockBenchmark.hybridClockNowSingleThread     thrpt   10  253800.841 ± 1555.133  ops/s
HybridClockBenchmark.hybridClockNowTenThreads       thrpt   10   35570.363 ±  581.248  ops/s
HybridClockBenchmark.hybridClockUpdateFiveThreads   thrpt   10   56725.796 ± 1055.919  ops/s
HybridClockBenchmark.hybridClockUpdateSingleThread  thrpt   10  241928.824 ± 3441.524  ops/s
HybridClockBenchmark.hybridClockUpdateTenThreads    thrpt   10   53343.884 ±  136.487  ops/s
```

https://issues.apache.org/jira/browse/IGNITE-23707