@vldpyatkov 2024-11-29
> Doesn't such an update use CAS under the hood? According to the code of `AtomicLongFieldUpdater`, it does:

In my opinion, incrementAndGet has better performance because it does not require a hold in an external loop. Instead of it, we just increment the internal value in the atomic variable.
We already compare this in micro-benchmark.